% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_nosiega.R
\name{gg_nosiega}
\alias{gg_nosiega}
\title{Presence-abscence ggplot for no cut area}
\usage{
gg_nosiega(dat)
}
\arguments{
\item{dat}{dataframe created with ind_nosiega()}
}
\value{
ggplot de ind_nosiega()
}
\description{
Presence-abscence ggplot for no cut area
}
\examples{
gg_nosiega(dat = ind_nosiega1)
}
