% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_siega.R
\name{gg_siega}
\alias{gg_siega}
\title{Presence-abscence ggplot for cut area}
\usage{
gg_siega(dat)
}
\arguments{
\item{dat}{dataframe created with ind_siega()}
}
\value{
ggplot de ind_siega()
}
\description{
Presence-abscence ggplot for cut area
}
\examples{
gg_siega(dat = ind_siega())
}
