% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_nosiega.R
\name{ind_nosiega}
\alias{ind_nosiega}
\title{Presence-abscence dataframe for no cut area}
\usage{
ind_nosiega(x)
}
\arguments{
\item{x}{empty}
}
\value{
random dataframe
}
\description{
Presence-abscence dataframe for no cut area
}
\examples{
ind_nosiega()
}
