% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_siega.R
\name{ind_siega}
\alias{ind_siega}
\title{Presence-abscence dataframe for cut area}
\usage{
ind_siega(x)
}
\arguments{
\item{x}{empty}
}
\value{
random dataframe
}
\description{
Presence-abscence dataframe for cut area
}
\examples{
ind_siega
}
